Replacement
Removed confusing separate folder and replaced old "master" code with
branch code.
